requirements.txt
alpaca-py
python-dotenv
yfinance
google-generativeai